[DEV-9734a] rename file
