+++ packages/chart/index.html
@@ -101,10 +101,10 @@
     <!-- <div class="react-container" data-config="/https://www.cdc.gov/poxvirus/mpox/modules/data-viz/mpx-age-sex1.json"></div> -->
 
     <!-- TESTS DATE EXCLUSIONS -->
-    <div
+    <!-- <div
       class="react-container"
       data-config="/examples/feature/tests-date-exclusions/date-exclusions-config.json"
-    ></div>
+    ></div> -->
     <!-- <div class="react-container" data-config="/examples/feature/tests-case-rate/case-rate-example-config.json"></div> -->
 
     <!-- TESTS BIG SMALL-->
