+++ packages/chart/src/_stories/ChartLine.Suppression.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import Chart from '../CdcChart.Component'
+import Chart from '../CdcChartComponent'
 import suppressionConfig from './_mock/suppression_mock.json'
 import SuppressedConfig from './_mock/bar-chart-suppressed.json'
 const meta: Meta<typeof Chart> = {
