+++ packages/dashboard/package.json
