+++ packages/dashboard/src/CdcDashboardComponent.tsx
@@ -19,7 +19,7 @@ import Loading from '@cdc/core/components/Loading'
 import { DataTransform } from '@cdc/core/helpers/DataTransform'
 import getViewport from '@cdc/core/helpers/getViewport'
 
-import CdcChart from '@cdc/chart/src/CdcChart'
+import CdcChart from '@cdc/chart/src/CdcChartComponent'
 import CdcDataBite from '@cdc/data-bite/src/CdcDataBite'
 import CdcMap from '@cdc/map/src/CdcMap'
 import CdcWaffleChart from '@cdc/waffle-chart/src/CdcWaffleChart'
@@ -353,6 +353,7 @@ export default function CdcDashboard({ initialState, isEditor = false, isDebug =
               <>
                 <Header visualizationKey={visualizationKey} subEditor='Chart' />
                 <CdcChart
+                  data={visualizationConfig.data}
                   key={visualizationKey}
                   config={visualizationConfig}
                   isEditor={true}
